SegmentEvaluator that turns SegmentLogic into an Ecto.Query.dynamic
This is pretty nifty because there's not that much code to represent any boolean logic statement.

The only thing that I don't really like about this is that gen_criteria is defined for each `object` type
separately. I wish there was a clean way to have a single function based on it, but the macro nature
of these functions makes some things a bit harder to do dynamically.